requirements.txt
boto3>=1.34.0
botocore>=1.34.0
sentence-transformers>=2.2.0
scikit-learn>=1.3.0
numpy>=1.24.0
requests>=2.31.0
python-dotenv>=1.0.0
pydantic>=2.0.0
typing-extensions>=4.7.0
langgraph>=0.0.40
langchain>=0.1.0
langchain-core>=0.1.0
html2text>=2020.1.16
pandas>=1.5.0
openpyxl>=3.0.0
streamlit>=1.28.0
pymongo